ADD equals and hashCode

+++ Online-Store/src/main/java/com/store/Online/Store/entity/Role.java
@@ -19,7 +19,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private int roleId;
+    private Long roleId;
 
     @Column(name = "role_name")
     private String roleName;
@@ -32,7 +32,7 @@ public class Role {
         this.roleName=roleName;
     }
 
-    public Role(int id,String roleName){
+    public Role(Long id,String roleName){
         this.roleName=roleName;
         this.roleId=id;
     }
